Implemented proposal_hash support
The proposal_hash is optional but if present, it validated to ensure the
vote is being cast on the right current version of the proposition.

To remain backward compatible, the proposal_hash remains a string on the
vote action but must be a valid sha256 format hexadecimal string.

The received hash, if non empty, is then compared against the current
proposal hash and we ensure both matches.